my_project/pyproject.toml
[tool.poetry]
name = "imdb_dataset"
version = "0.1.0"
description = "The IMDB_Dataset API project aims to seamlessly empower users with access to the dynamic world of the IMDB dataset."
authors = ["JunyinLuo <jl6530@columbia.edu>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
beautifulsoup4 = "^4.12.2"
bs4 = "^0.0.1"
h5py = "^3.9.0"
html5lib = "^1.1"
importlib-metadata = "^6.8.0"
iniconfig = "^2.0.0"
lxml = "^4.9.3"
matplotlib = "^3.5.3"
packaging = "^23.1"
pandas = "^2.1.1"
pandas-datareader = "^0.10.0"
Pillow = "^10.0.1"
pyarrow = "^14.0.1"
pyparsing = "^3.1.1"
pytest = "^7.4.3"
pytz = "^2023.3.post1"
requests = "^2.31.0"
requests-oauthlib = "^1.3.1"
urllib3 = "^2.0.5"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
